// packages/@fluentui-vue/next/src/components/Facepile/Facepile.types.ts
export interface IFacepilePersona {
  /**
    * Name of the person.
    */
  personaName?: string;

  /**
    * Url to the image to use, should be a square aspect ratio and big enough to fit in the image area.
    */
  imageUrl?: string;

  /**
    * The user's initials to display in the image area when there is no image.
    * @defaultvalue Derived from `personaName`
    */
  imageInitials?: string;

  /**
    * Whether initials are calculated for phone numbers and number sequences.
    * Example: Set property to true to get initials for project names consisting of numbers only.
    * @defaultvalue false
    */
  allowPhoneInitials?: boolean;

  /**
    * The background color when the user's initials are displayed.
    * @defaultvalue Derived from `personaName`
    */
  initialsColor?: PersonaInitialsColor;

  /**
    * If provided, persona will be rendered with cursor:pointer and the handler will be
    * called on click.
    */
  onClick?: (ev?: MouseEvent, persona?: IFacepilePersona) => void;

  /**
    * If provided, the handler will be called on mouse move.
    */
  onMouseMove?: (ev?: MouseEvent, persona?: IFacepilePersona) => void;

  /**
    * If provided, the handler will be called when mouse moves out of the component.
    */
  onMouseOut?: (ev?: MouseEvent, persona?: IFacepilePersona) => void;

  /**
    * Extra data - not used directly but can be handy for passing additional data to custom event
    * handlers.
    */
  data?: any;

  /**
    * Optional keytip for this button that is only added when 'onClick' is defined for the persona
    */
  keytipProps?: IKeytipProps;
}

/**
  * {@docCategory Facepile}
  */
export enum OverflowButtonType {
  /** No overflow */
  none = 0,
  /** +1 overflow icon */
  descriptive = 1,
  /** More overflow icon */
  more = 2,
  /** Chevron overflow icon */
  downArrow = 3,
}

/**
  * {@docCategory Facepile}
  */
export interface IFacepileStyleProps {
  /**
    * Theme provided by High-Order Component.
    */
  theme: ITheme;

  /**
    * Accept custom classNames
    */
  className?: string;

  /**
    * Pixel value for spacing around button. Number value set in pixels
    */
  spacingAroundItemButton?: number;
}

/**
  * {@docCategory Facepile}
  */
export interface IFacepileStyles {
  /**
    * Style for the root element.
    */
  root: IStyle;
  addButton: IStyle;
  descriptiveOverflowButton: IStyle;
  itemContainer: IStyle;
  itemButton: IStyle;
  members: IStyle;
  member: IStyle;
  overflowButton: IStyle;
  overflowInitialsIcon: IStyle;
  screenReaderOnly: IStyle;
}